Module smash::app::lua_bind::AttackModule

Functions

add_power⚠

attack_data⚠

attack_direction⚠

attack_part_speed⚠

attack_reference_pos⚠

center_pos⚠

clear⚠

clear_all⚠

clear_inflict_kind_status⚠

clear_lerp⚠

customize_attack_ratio⚠

damage_effect_mul_scale⚠

damage_shake_scale⚠

disable_tip⚠

enable_attack_r_fix_damage_speed_up_all⚠

enable_safe_pos⚠

get_attack_power_mul_pattern⚠

get_attack_power_mul_scale⚠

get_attacker_attribute⚠

get_critical_attack_damage_mul⚠

get_critical_attack_reaction_mul⚠

get_damage_reaction_mul⚠

get_inflict⚠

get_inflict_status⚠

get_lr⚠

get_offset2⚠

get_power⚠

get_power_apply_defense⚠

get_power_up⚠

get_power_up_attr⚠

group⚠

hit_absolute⚠

hit_absolute_joint⚠

indirect_object_id⚠

indirect_team_no⚠

init_attack_pos⚠

is_attack⚠

is_attack_occur⚠

is_constraint_pos⚠

is_critical_attack⚠

is_hit_abs⚠

is_infliction⚠
Returns whether or not you are currently hitting an opponent (in hitlag).

is_infliction_status⚠
Returns whether or not you have hit something during the current status

is_invalid_invincible⚠

is_invalid_xlu⚠

is_no_dead_damage_fly_effect⚠

is_power_up_reaction⚠

lr_check_default⚠

off_target_kind⚠

part_size⚠

pos_x⚠

pos_x_2⚠

pos_y⚠

power_add_status⚠

power_mul⚠

power_mul_2nd⚠

power_mul_3rd⚠

power_mul_4th⚠

power_mul_5th⚠

power_mul_status⚠

power_speed_limit⚠

power_speed_status⚠

reaction_add⚠

reaction_effect⚠

reaction_fix⚠

reaction_mul⚠

reaction_mul_2nd⚠

reaction_mul_3rd⚠

reaction_mul_4th⚠

ref_log_group⚠

relocate⚠

reset_status_attack⚠

set_absolute⚠

set_absolute_lr⚠

set_accept_no_lr⚠

set_add_reaction_frame⚠

set_add_reaction_frame_revised⚠

set_attack⚠

set_attack_2⚠

set_attack_camera_quake⚠

set_attack_camera_quake_forced⚠

set_attack_composition_speed⚠

set_attack_height⚠

set_attack_height_all⚠

set_attack_keep_rumble⚠

set_attack_level⚠

set_attack_no_reaction_search⚠

set_attack_no_weight_reaction⚠

set_attack_no_weight_reaction_all⚠

set_attack_power_mul_pattern⚠

set_attack_power_mul_scale⚠

set_attack_reference_joint_id⚠

set_attack_scale⚠

set_attack_scale_2nd⚠

set_attacker_attribute⚠

set_base_offset⚠

set_captured_same_time_attack⚠

set_captured_same_time_attack_damage_mul⚠

set_catch_only⚠

set_catch_only_all⚠

set_constraint_pos⚠

set_critical_attack_damage_mul⚠

set_critical_attack_reaction_mul⚠

set_customize_attack_ratio⚠

set_damage_effect_mul_scale⚠

set_damage_reaction_mul⚠

set_damage_shake_scale⚠

set_disable_power_add_status_zero⚠

set_final_finish_cut_in⚠

set_force_reaction⚠

set_ice_frame_mul⚠

set_ignore_just_shield⚠

set_indirect_info⚠

set_ink_value⚠

set_invalid_invincible⚠

set_invalid_xlu⚠

set_is_critical_attack⚠

set_latest_absolute⚠

set_latest_no⚠

set_lerp⚠

set_lerp_ratio⚠

set_lr_check_back⚠

set_lr_check_default⚠

set_lr_check_front⚠

set_lr_check_front_lr⚠

set_nearest⚠

set_no_attacker_log⚠

set_no_comp_damage_motion⚠

set_no_damage_fly_smoke_all⚠

set_no_damage_orbit_all⚠

set_no_dead_all⚠

set_no_dead_damage_fly_effect⚠

set_no_finish_camera⚠

set_no_finish_camera_ex⚠

set_no_hop_opponent⚠

set_no_hop_opponent_all⚠

set_no_squat_damage_reaction_mul⚠

set_no_team⚠

set_no_uniq_effect_all⚠

set_no_world_scale⚠

set_node⚠

set_offset⚠

set_offset2⚠

set_optional_hit_effect⚠

set_optional_hit_sound⚠

set_overlap_hit⚠

set_paralyze_frame⚠

set_poison_param⚠

set_pos_x⚠

set_power⚠

set_power_add_status⚠

set_power_mul⚠

set_power_mul_2nd⚠

set_power_mul_3rd⚠

set_power_mul_4th⚠

set_power_mul_5th⚠

set_power_mul_status⚠

set_power_speed_limit⚠

set_power_speed_status⚠

set_power_up⚠

set_power_up_attr⚠

set_power_up_reaction⚠

set_reaction_add⚠

set_reaction_effect⚠

set_reaction_fix⚠

set_reaction_mul⚠

set_reaction_mul_2nd⚠

set_reaction_mul_3rd⚠

set_reaction_mul_4th⚠

set_reflect_attack⚠

set_restrict⚠

set_serial_hit_frame⚠

set_set_off⚠

set_shield_stiff_mul⚠

set_size⚠

set_special_paint⚠

set_speed⚠

set_target_category⚠

set_vec_target_pos⚠

set_vector⚠

set_whole⚠

shield_stiff_mul⚠

size⚠

sleep⚠

sleep_partialy⚠

speed⚠

speed_x⚠

un_ref_log_group⚠