Module smash::app::lua_bind::ControlModule

Functions

add_clatter_time⚠

check_button_off⚠
Returns whether or not the current module_accessor is NOT holding the specified CONTROL_PAD_BUTTON on the same frame this function was run

check_button_on⚠
Returns whether or not the current module_accessor is holding the specified CONTROL_PAD_BUTTON on the same frame this function was run

check_button_on_release⚠

check_button_on_trriger⚠

check_button_release⚠

check_button_trigger⚠
Returns whether or not the current module_accessor has pressed the specified CONTROL_PAD_BUTTON on the same frame this function was run

clear_command⚠

clear_command_one⚠

clear_rumble_hit_data⚠

delete_reserve_turn_lr⚠

end_clatter⚠

end_clatter_motion_rate⚠

exec_command⚠

get_attack_air_kind⚠

get_attack_air_stick_dir⚠

get_attack_air_stick_x⚠

get_attack_air_stick_y⚠

get_attack_hi3_fb_kind⚠

get_attack_lw3_fb_kind⚠

get_button⚠

get_button_prev⚠

get_clatter_threshold⚠

get_clatter_time⚠

get_command_flag_cat⚠
returns the current "command flag" based on the player's current controller inputs. Command flags come in 4 different categories.

get_command_life⚠

get_command_life_count_max⚠

get_down_stand_fb_kind⚠

get_flick_after_x⚠

get_flick_after_x_dir⚠

get_flick_after_y⚠

get_flick_bonus⚠

get_flick_bonus_lr⚠

get_flick_no_reset_x⚠

get_flick_no_reset_y⚠

get_flick_sub_x⚠

get_flick_sub_x_dir⚠

get_flick_sub_y⚠

get_flick_sub_y_dir⚠

get_flick_x⚠

get_flick_x_dir⚠

get_flick_y⚠

get_flick_y_dir⚠

get_lr⚠

get_pad_flag⚠

get_release⚠

get_reserve_turn_lr⚠

get_special_command_lr⚠

get_stick_angle⚠

get_stick_dir⚠

get_stick_prev_x⚠

get_stick_prev_y⚠

get_stick_x⚠
Returns the current x-axis position of the main control stick in a -1 to 1 range. (I.E. -1 = all the way left, 1 = all the way right, 0 = neutral)

get_stick_x_no_clamp⚠

get_stick_y⚠
Returns the current y-axis position of the main control stick in a -1 to 1 range. (I.E. -1 = all the way down, 1 = all the way up, 0 = neutral)

get_stick_y_no_clamp⚠

get_sub_stick_dir⚠

get_sub_stick_prev_x⚠

get_sub_stick_prev_y⚠

get_sub_stick_x⚠

get_sub_stick_y⚠

get_trigger⚠

get_trigger_count⚠

get_trigger_count_prev⚠

is_clatter_stop⚠

is_clear_command⚠

is_enable_flick_jump⚠

is_input_clatter⚠

is_jump_mini_button⚠

is_remake_command⚠

is_stick_reversed⚠

is_stick_side⚠

is_sub_stickSide⚠

item_heavy_throw_fb_kind⚠

item_light_throw_air_fb4_kind⚠

item_light_throw_air_fb_kind⚠

item_light_throw_fb4_kind⚠

item_light_throw_fb_kind⚠

request_rumble_hit⚠

reset_attack_air_kind⚠

reset_button⚠

reset_down_stand_fb_kind⚠

reset_flick_bonus⚠

reset_flick_bonus_lr⚠

reset_flick_sub_x⚠

reset_flick_sub_y⚠

reset_flick_x⚠

reset_flick_y⚠

reset_main_stick⚠

reset_main_stick_x⚠

reset_main_stick_y⚠

reset_special_command⚠

reset_sub_stick⚠

reset_sub_stick_x⚠

reset_sub_stick_y⚠

reset_trigger⚠

reset_trigger_2⚠

reset_turn_lr⚠

reverse_special_command⚠

reverse_x_frame⚠

set_add_jump_mini_button_life⚠

set_attack_air_kind⚠

set_back_command⚠

set_clatter_shake_scale⚠

set_clatter_stop⚠

set_clatter_time⚠

set_command_life_extend⚠

set_dec_time⚠

set_dec_time_recovery⚠

set_down_stand_fb_kind⚠

set_log_active⚠

set_main_stick_x⚠

set_main_stick_y⚠

set_off⚠

set_reverse_x_frame⚠

set_rumble⚠

set_rumble_all⚠

set_rumble_hit_data⚠

set_special_command⚠

set_special_command_life_count_extend⚠

set_special_command_life_extend⚠

set_stick_reverse⚠

special_s_turn⚠

start_clatter⚠

start_clatter_motion_rate⚠

stop_rumble⚠

stop_rumble_all⚠

stop_rumble_id⚠

stop_rumble_kind⚠