Module smash::app::lua_bind::GroundModule

Functions

attach⚠

attach_ground⚠

can_entry_cliff⚠

can_entry_cliff_hang_data⚠

can_entry_cliff_same⚠

center_pos⚠

check_down_correct_pos⚠

clear_cliff_point⚠

clear_pass_floor⚠

cliff_check⚠

correct⚠

correct_pos⚠

detach⚠

entry_cliff⚠

get_center_pos⚠
Returns the X position of the base of your ECB relative the the origin of the current stage.

get_circle_radius⚠

get_cliff_id_uint32⚠

get_cliff_movement_speed⚠

get_correct⚠

get_correct_pos⚠

get_correct_pos_local⚠

get_distance_to_floor⚠

get_down_correct_edge_pos⚠

get_down_friction⚠

get_down_movement_speed⚠

get_down_pos⚠

get_latest_down_correct_pos⚠

get_left_pos⚠

get_line_movement_speed⚠

get_offset_x⚠
returns the x offset for the current battle object's ECB's base. The base of the ECB rhombus is between the character's feet

get_offset_y⚠
returns the y offset for the current battle object's ECB's base. The base of the ECB rhombus is between the character's feet

get_rhombus⚠

get_right_pos⚠

get_shape_kind⚠

get_shape_safe_pos⚠

get_touch_flag⚠

get_touch_line_raw⚠

get_touch_material_type⚠

get_touch_moment_flag⚠

get_touch_normal⚠

get_touch_normal_consider_gravity⚠

get_touch_normal_for_touch_attack_data⚠

get_touch_normal_x⚠

get_touch_normal_x_consider_gravity⚠

get_touch_normal_y⚠

get_touch_normal_y_consider_gravity⚠

get_touch_pos⚠

get_touch_wall_cliff_pos⚠

get_up_pos⚠

get_width⚠

get_z⚠

ground_touch_flag⚠

ground_touch_flag_ex⚠

hang_can_entry_cliff_dir⚠

hang_can_entry_cliff_pos⚠

hang_cliff_dir⚠

hang_cliff_pos⚠

hang_cliff_pos_3f⚠
returns the coordinates of the nearest cliff as a Vector3f

is_attach⚠

is_attach_cliff⚠

is_attachable⚠

is_cliff_move_exception⚠

is_floor_touch_line⚠

is_floor_vanishing⚠

is_ignore_fighter_other⚠

is_ignore_fighter_other_wall_l⚠

is_ignore_fighter_other_wall_r⚠

is_miss_foot⚠

is_myground_nearcliff⚠

is_near_cliff⚠

is_ottotto⚠

is_ottotto_lr⚠

is_passable_check⚠

is_passable_ground⚠

is_status_cliff⚠

is_still_on_floor⚠

is_touch⚠

is_wall_touch_line⚠

leave_cliff⚠

line_segment_check⚠

modify_rhombus⚠

pass_floor⚠

ray_check⚠

ray_check_get_line⚠

ray_check_get_line_hit_pos⚠

ray_check_get_line_hit_pos_ignore_any⚠

ray_check_get_line_hit_pos_ignore_any_no_culling⚠

ray_check_get_line_hit_pos_no_culling⚠

ray_check_get_line_ignore_any⚠

ray_check_get_line_ignore_any_no_culling⚠

ray_check_get_line_no_culling⚠

ray_check_get_line_target_any⚠

ray_check_get_line_target_any_no_culling⚠

ray_check_hit_normal⚠

ray_check_hit_pos⚠

ray_check_hit_pos_normal⚠

ray_check_hit_pos_normal_no_culling⚠

ray_check_hit_pos_target⚠

reentry_cliff⚠

select_cliff_hangdata⚠

set_attach_ground⚠

set_auto_detach⚠

set_cliff_check⚠

set_cloud_through⚠

set_coll_stop_slidable_length⚠

set_collidable⚠

set_correct⚠

set_correct_ignore_slope⚠

set_gr_collision_mode⚠

set_ignore_boss⚠

set_ignore_fighter_other⚠

set_ignore_friction⚠

set_ignore_line_type1⚠

set_ignore_slide_up⚠

set_init_circle⚠

set_init_shape_kind⚠

set_keep_distant_cliff⚠

set_no_cliff_stop_energy⚠

set_offset_x⚠

set_offset_y⚠

set_passable_check⚠

set_reverse_direction⚠

set_rhombus_modify⚠

set_rhombus_modify_air_lasso⚠

set_rhombus_offset⚠
Sets the offset for the current battle object's ECB (Environmental collision box)

set_shape_data_rhombus_modify_node_offset⚠

set_shape_flag⚠

set_shape_safe_pos⚠

set_status_ground⚠

set_test_coll_stop⚠

set_test_coll_stop_status⚠

set_update_shape⚠

set_z⚠

test⚠

test_ground⚠

update_force⚠

update_shape⚠