Function smash::app::lua_bind::GroundModule::hang_cliff_pos_3f

pub unsafe extern "C" fn hang_cliff_pos_3f(
    arg1: *mut BattleObjectModuleAccessor
) -> Vector3f

returns the coordinates of the nearest cliff as a Vector3f

Arguments

module_accessor - Pointer to BattleObjectModuleAccessor

Example

// store the position of the current nearest cliff if the player is in the ledge hang status
let mut pos: smash::phx::Vector3f = smash::phx::Vector3f { x: 0.0, y: 0.0, z: 0.0 };
if StatusModule::status_kind(module_accessor) == *FIGHTER_STATUS_KIND_CLIFF_WAIT {
    pos = GroundModule::hang_cliff_pos_3f(module_accessor);
}