Module smash::app::lua_bind::KineticModule

Functions

add_speed⚠
Adds the speed specified by the Vector3f to the battle object's current speed

add_speed_outside⚠

change_kinetic⚠
Changes the current KINETIC_TYPE to the specified value

clear_speed_all⚠

clear_speed_attr⚠

clear_speed_energy_id⚠

enable_energy⚠

get_energy⚠

get_kinetic_type⚠
Returns the current KINETIC_TYPE

get_sum_rotation⚠

get_sum_speed⚠

get_sum_speed3f⚠
Returns the current velocity based on the specified kinetic energy attribute as a Vector3f

get_sum_speed_length⚠

get_sum_speed_x⚠
Returns the current x velocity based on the specified kinetic energy attribute

get_sum_speed_y⚠
Returns the current y velocity based on the specified kinetic energy attribute

is_enable_energy⚠

is_suspend_energy⚠

mul_accel⚠
multiplies the current (specified type of) acceleration by the provided vector3f

mul_speed⚠
multiplies the current (specified type of) speed by the provided vector3f

reflect_accel⚠

reflect_speed⚠

resume_energy⚠

resume_energy_all⚠

set_consider_ground_friction⚠

sleep⚠

suspend_energy⚠

suspend_energy_all⚠

unable_energy⚠

unable_energy_all⚠