Module smash::app::lua_bind::PostureModule

Functions

add_pos⚠

add_pos_2d⚠

base_scale⚠

init_pos⚠

init_rot⚠

init_rot_2⚠

init_rot_y_lr⚠

init_scale⚠

is_rot_y_lr_different_inner_lr⚠

lr⚠
Returns the current direction the current battle object is facing. -1 = left, 1 = right

owner_scale⚠

pos⚠

pos_2d⚠

pos_x⚠
Returns stage X position of the object

pos_y⚠
Returns stage Y position of the object

pos_z⚠
Returns stage Z position of the object

prev_pos⚠

prev_pos_2d⚠

reverse_lr⚠
Reverses the direction that the current battle object is facing

reverse_rot_y_lr⚠

rot⚠

rot_x⚠

rot_y⚠

rot_y_lr⚠

rot_z⚠

scale⚠

set_link_scale⚠

set_lr⚠
Sets the direction faced for the current battle object.

set_owner_scale⚠

set_pos⚠

set_pos_2d⚠

set_rot⚠

set_scale⚠

set_stick_lr⚠

set_sync_constraint_joint⚠

update_rot_y_lr⚠
Updates the current battle object's orientation. Usually used in conjunction with PostureModule::set_lr or PostureModule::reverse_lr